+++ src/kabusys/data/news_collector.py
@@ -24,10 +24,8 @@ from __future__ import annotations
 
 import gzip
 import hashlib
-import ipaddress
 import logging
 import re
-import socket
 import urllib.error
 import urllib.parse
 import urllib.request
@@ -38,6 +36,12 @@ from typing import TypedDict
 import duckdb
 from defusedxml import ElementTree as ET
 from defusedxml.common import DefusedXmlException
+
+from kabusys.utils.http import (
+    SSRFBlockRedirectHandler as _SSRFBlockRedirectHandler,
+    is_private_host as _is_private_host,
+    validate_url_scheme as _validate_url_scheme,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -128,55 +132,6 @@ def _make_article_id(url: str) -> str:
     return hashlib.sha256(_normalize_url(url).encode()).hexdigest()[:32]
 
 
-def _validate_url_scheme(url: str) -> None:
-    """URL のスキームが http または https であることを検証する。
-
-    SSRF / ローカルファイル読み出しを防ぐため、http/https 以外を拒否する。
-
-    Args:
-        url: 検証対象の URL。
-
-    Raises:
-        ValueError: スキームが http/https でない場合。
-    """
-    scheme = urllib.parse.urlparse(url).scheme.lower()
-    if scheme not in ("http", "https"):
-        raise ValueError(f"許可されていないURLスキーム: {scheme!r} (url={url!r})")
-
-
-def _is_private_host(hostname: str | None) -> bool:
-    """ホスト/IP がプライベート・ループバック・リンクローカルかを判定する。
-
-    SSRF 対策としてリダイレクト先の内部アドレスへのアクセスを拒否するために使用する。
-    IP アドレスは直接判定し、ホスト名は DNS 解決して判定する。
-    DNS 解決失敗時は安全側（非プライベート）とみなす。
-
-    Args:
-        hostname: 検査対象のホスト名または IP アドレス文字列。
-
-    Returns:
-        プライベート/ループバック/リンクローカル/マルチキャストの場合 True。
-    """
-    if not hostname:
-        return True
-    # IP アドレスとして直接解析できる場合
-    try:
-        ip = ipaddress.ip_address(hostname)
-        return ip.is_private or ip.is_loopback or ip.is_link_local or ip.is_multicast
-    except ValueError:
-        pass
-    # ホスト名の場合: DNS 解決して全 A/AAAA レコードを検査
-    try:
-        for info in socket.getaddrinfo(hostname, None, proto=socket.IPPROTO_TCP):
-            ip = ipaddress.ip_address(info[4][0])
-            if ip.is_private or ip.is_loopback or ip.is_link_local or ip.is_multicast:
-                return True
-    except (OSError, ValueError):
-        # 解決失敗は非プライベートとみなして通過させる
-        pass
-    return False
-
-
 def preprocess_text(text: str | None) -> str:
     """テキストの前処理を行う。
 
@@ -226,23 +181,6 @@ def _parse_rss_datetime(date_str: str | None) -> datetime:
 # ---------------------------------------------------------------------------
 # RSS 取得
 # ---------------------------------------------------------------------------
-
-
-class _SSRFBlockRedirectHandler(urllib.request.HTTPRedirectHandler):
-    """リダイレクト時にスキームとプライベートアドレスを事前検証するハンドラ。
-
-    接続前にリダイレクト先を検査することで、内部ネットワークへの到達を防ぐ。
-    """
-
-    def redirect_request(self, req, fp, code, msg, headers, newurl):  # type: ignore[override]
-        parsed = urllib.parse.urlparse(newurl)
-        if parsed.scheme.lower() not in ("http", "https"):
-            raise urllib.error.URLError(f"リダイレクト先のスキームが不正: {newurl!r}")
-        if _is_private_host(parsed.hostname):
-            raise urllib.error.URLError(
-                f"リダイレクト先がプライベートアドレス: {newurl!r}"
-            )
-        return super().redirect_request(req, fp, code, msg, headers, newurl)
 
 
 def _urlopen(req: urllib.request.Request, timeout: int):
